+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## master
 
+- CHANGED: Rescue all SystemCallError instead of a few Errno errors (GH-212). [Thanks @mat813]
+
 - FIXED: whois.ua should support status `ok`.
 
 - FIXED: whois.registry.net.za crashes when the registrant doesn't have a name.