+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## master
 
+- NEW: Added #domain and #registrar to whois.dns.be parser (GH-216). [Thanks @chuckadams]
+
 - CHANGED: Rescue all SystemCallError instead of a few Errno errors (GH-212). [Thanks @mat813]
 
 - FIXED: whois.ua should support status `ok`.